# fastapi_backend/openapi.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "北京健康评估系统API",
    "description": "北京健康评估系统的后端API接口",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/users/": {
      "post": {
        "tags": ["用户管理"],
        "summary": "创建新用户",
        "operationId": "create_user",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "用户创建成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "用户名已存在"
          },
          "401": {
            "description": "未经授权"
          },
          "403": {
            "description": "权限不足"
          }
        }
      },
      "get": {
        "tags": ["用户管理"],
        "summary": "获取用户列表",
        "operationId": "read_users",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "description": "跳过的记录数",
            "schema": {
              "type": "integer",
              "default": 0
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "返回的最大记录数",
            "schema": {
              "type": "integer",
              "default": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功获取用户列表",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          },
          "401": {
            "description": "未经授权"
          },
          "403": {
            "description": "权限不足"
          }
        }
      }
    },
    "/api/users/me": {
      "get": {
        "tags": ["用户管理"],
        "summary": "获取当前用户信息",
        "operationId": "read_user_me",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功获取当前用户信息",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "401": {
            "description": "未经授权"
          }
        }
      }
    },
    "/api/users/{user_id}": {
      "get": {
        "tags": ["用户管理"],
        "summary": "获取特定用户信息",
        "operationId": "read_user",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功获取用户信息",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "401": {
            "description": "未经授权"
          },
          "403": {
            "description": "权限不足"
          },
          "404": {
            "description": "用户不存在"
          }
        }
      },
      "put": {
        "tags": ["用户管理"],
        "summary": "更新用户信息",
        "operationId": "update_user",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "成功更新用户信息",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "用户名已存在"
          },
          "401": {
            "description": "未经授权"
          },
          "403": {
            "description": "权限不足"
          },
          "404": {
            "description": "用户不存在"
          }
        }
      },
      "delete": {
        "tags": ["用户管理"],
        "summary": "删除用户",
        "operationId": "delete_user",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "成功删除用户"
          },
          "400": {
            "description": "不能删除当前登录的管理员账户"
          },
          "401": {
            "description": "未经授权"
          },
          "403": {
            "description": "权限不足"
          },
          "404": {
            "description": "用户不存在"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "fullname": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "user_type": {
            "type": "integer"
          },
          "is_active": {
            "type": "boolean"
          },
          "create_time": {
            "type": "string",
            "format": "date-time"
          },
          "roles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Role"
            }
          }
        }
      },
      "UserCreate": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "fullname": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "user_type": {
            "type": "integer"
          },
          "is_active": {
            "type": "boolean",
            "default": true
          }
        },
        "required": ["username", "password", "fullname", "email", "user_type"]
      },
      "UserUpdate": {
        "type": "object",
        "properties": {
          "password": {
            "type": "string"
          },
          "fullname": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "user_type": {
            "type": "integer"
          },
          "is_active": {
            "type": "boolean"
          }
        }
      },
      "Role": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "permissions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Permission"
            }
          }
        }
      },
...
      "Permission": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
} 